add token, description for methods, refactoring
