+++ plugins/de.cognicrypt.staticanalyzer/resources/Token.txt
@@ -1,0 +1,1 @@
+MzY5OTllNTI1ZjU3Y2UwYzViYTZkNWYwZjQ1OTgzN2VmOThhNTk3OA
